Renaming of LabelDto and ProjectCollaborator folders
