+++ Todo.Api/Controllers/LabelsController.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Todo.Api.Interfaces;
 using Todo.Api.Mappers;
 using Todo.Api.Repositories.Interfaces;
-using Todo.Core.Dtos.LabelDto;
+using Todo.Core.Dtos.LabelDtos;
 
 namespace Todo.Api.Controllers;
