+++ Todo.Api/Mappers/ProjectCollaboratorMapper.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics;
 using Microsoft.OpenApi.Extensions;
-using Todo.Core.Dtos.ProjectCollaborator;
+using Todo.Core.Dtos.ProjectCollaboratorDtos;
 using Todo.Core.Entities;
 
 namespace Todo.Api.Mappers;
