+++ Todo.Core/Dtos/LabelDtos/UpdateLabelDto.cs
@@ -1,13 +1,14 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Todo.Core.Dtos.LabelDto;
+namespace Todo.Core.Dtos.LabelDtos;
 
-public class CreateLabelDto
+public class UpdateLabelDto
 {
-    public string Name { get; set; } = string.Empty;
+    public string? Name { get; set; } = null;
 
     [StringLength(7)]
     [RegularExpression("^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
         ErrorMessage = "Color must be a valid hex color code. Did you remember to include the '#'?")]
+
     public string? Color { get; set; } = null;
 }
