+++ Todo.Core/Dtos/ProjectCollaboratorDtos/ProjectCollaboratorDto.cs
@@ -1,4 +1,4 @@
-namespace Todo.Core.Dtos.ProjectCollaborator;
+namespace Todo.Core.Dtos.ProjectCollaboratorDtos;
 
 public class ProjectCollaboratorDto
 {
